+++ src/global/composeRefs.ts
@@ -1,6 +1,5 @@
 import {Ref, MutableRefObject} from 'react';
 import memoizeOne from 'memoize-one';
-import deprecate from 'util-deprecate';
 
 function composeRefs<T>(...refs: (Ref<T> | undefined)[]) {
   return (value: T | null) => refs.forEach(ref => {
@@ -11,9 +10,6 @@ function composeRefs<T>(...refs: (Ref<T> | undefined)[]) {
     }
   });
 }
-
-// TODO remove export in 7.0, composeRefs should be used only in createComposedRef and in useComposedRefs in the future
-export default deprecate(composeRefs, 'composeRefs is deprecated and will be removed in 7.0. Use createComposedRef instead.');
 
 export function createComposedRef<T>() {
   return memoizeOne(composeRefs<T>);