+++ src/header/tray-icon.tsx
@@ -1,3 +1,0 @@
-// TODO remove this file in 7.0
-
-export {type HeaderIconSpecificProps as TrayIconSpecificProps, default} from './header-icon';